requirements.txt
openai
python-dotenv
rich
requests